Fix the calculation logic
Each part of the calculation formula has its own function that checks that the correct characters have been entered into the calculator.

When the calculator is ready to calculate, it selects the selected calculation according to the operator.

Finally, the calculator clears the given values ​​or changes them, depending on the selected process.