Add logic for decimal numbers
An object was created from each step of the calculation operation.
This made activating the different calculation step easier, as well as turning off the decimal button.
The logic also had to create numbers that were too big, and prevent division by zero.

In the HTML file for buttons, separate classes were created to make it easier to call numbers and operators separately.